ADD id's DB for API DATA INSERTS + README.TXT FOR DB

+++ UTNAppDB/Scripts UTNAppDB/Programability/Company SP's/CREATE_InsertCompany_SP.sql
@@ -10,7 +10,7 @@ CREATE PROCEDURE InsertCompany
 	IN Sector varchar(200), 
 	IN Name varchar(200), 
 	IN Description varchar(200), 
-	IN Cuit int, 
+	IN Cuit bigint, 
 	IN CompanyLink varchar(200), 
 	IN AboutUs varchar(200)
